mybatis-plus downgrade
3.5.2 -> 3.4.2

+++ cloud-comment-service-provider12500/src/main/java/usyd/mingyi/springcloud/controller/CommentController.java
@@ -42,7 +42,7 @@ public class CommentController {
     public R<Long> countCommentsReceived(){
         Long userId = BaseContext.getCurrentId();
 
-        Long count = commentService.count(new LambdaQueryWrapper<Comment>()
+        Long count = (long) commentService.count(new LambdaQueryWrapper<Comment>()
                 .eq(Comment::getTargetUserId, userId)
                 .eq(Comment::getIsRead,false));
         return  R.success(count);
