+++ cloud-interaction-service-provider12600/src/main/java/usyd/mingyi/springcloud/service/MentionServiceImp.java
@@ -44,7 +44,7 @@ public class MentionServiceImp extends ServiceImpl<MentionMapper, Mention> imple
     @Override
     public Long countMentionsReceived(Long userId) {
         LambdaQueryWrapper<Mention> query = new LambdaQueryWrapper<>();
-        return this.count(
+        return (long) this.count(
                 query.eq(Mention::getTargetUserId,userId)
                 .eq(Mention::getIsRead, false)
         );
